client/requirements.txt
flask
flask[async]
flask-bootstrap4
flask-wtf
pyotp
python-dotenv
websockets